sql datatypes improvements

+++ database/migrations/2019_09_17_165224_category_product.php
@@ -14,7 +14,7 @@ class CategoryProduct extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('product_id');
             $table->unsignedSmallInteger('category_id');
         });
     }
