+++ database/migrations/2019_11_08_134428_create_mas_orders_table.php
@@ -14,8 +14,8 @@ class CreateMasOrdersTable extends Migration
     public function up()
     {
         Schema::create('mas_orders', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('order_id');
+            $table->increments('id');
+            $table->unsignedInteger('order_id');
 
             $table->string('mas_control_number')->nullable();
             $table->string('mas_message_number')->nullable();
